requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
pillow
seaborn
streamlit
opencv-python